Update README.md
This project is a custom remote that controls smart lights and Wemos devices around a room. It uses buttons to turn devices on and off, switch between different lights and Wemos, change light colors, and adjust brightness. A photodiode controls Neopixels based on light in the room, and an OLED screen shows the current settings. I made this because I enjoy using remotes to control lights and wanted to build my own version that works with LED strips and smart lights.